created function to create has of configmap file
